ci: Update pre-commit plugins and actions
I've been using conventional commits as much as possible for the past 12
months so adding in a pre-commit action to make it official. Also, most
of my GitHub repos have a protection on the main branch so try to
prevent accidental commits to the main branch. Finally, use a consistent
YAML style for list elements.